# restful-web-service/src/main/resources/swagger.yml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: 'urn:tos'
  contact: {}
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: 'localhost:8080'
basePath: /
tags:
  - name: Category
    description: Category Product Controller
  - name: Product
    description: Category Product Controller
paths:
  /categories:
    get:
      tags:
        - Category
        - Product
      summary: retrieveAllCategories
      operationId: retrieveAllCategoriesUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Category'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - Category
        - Product
      summary: createCategory
      operationId: createCategoryUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: category
          description: category
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/categories/{categoryId}/products':
    get:
      tags:
        - Category
        - Product
      summary: retrieveAllProducts
      operationId: retrieveAllProductsUsingGET
      produces:
        - '*/*'
      parameters:
        - name: categoryId
          in: path
          description: categoryId
          required: true
          type: integer
          format: int64
        - name: offset
          in: query
          required: false
          type: integer
          format: int64
        - name: pageNumber
          in: query
          required: false
          type: integer
          format: int32
        - name: pageSize
          in: query
          required: false
          type: integer
          format: int32
        - name: paged
          in: query
          required: false
          type: boolean
        - name: sort.sorted
          in: query
          required: false
          type: boolean
        - name: sort.unsorted
          in: query
          required: false
          type: boolean
        - name: unpaged
          in: query
          required: false
          type: boolean
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Product'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - Category
        - Product
      summary: createProduct
      operationId: createProductUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: categoryId
          in: path
          description: categoryId
          required: true
          type: integer
          format: int64
        - in: body
          name: product
          description: product
          required: true
          schema:
            $ref: '#/definitions/Product'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Product'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/categories/{categoryId}/products/{productId}':
    delete:
      tags:
        - Category
        - Product
      summary: deleteProduct
      operationId: deleteProductUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: categoryId
          in: path
          description: categoryId
          required: true
          type: integer
          format: int64
        - name: productId
          in: path
          description: productId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
    patch:
      tags:
        - Category
        - Product
      summary: updateProduct
      operationId: updateProductUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: categoryId
          in: path
          description: categoryId
          required: true
          type: integer
          format: int64
        - in: body
          name: newProduct
          description: newProduct
          required: true
          schema:
            $ref: '#/definitions/Product'
        - name: productId
          in: path
          description: productId
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Product'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  '/categories/{id}':
    get:
      tags:
        - Category
        - Product
      summary: retrieveCategory
      operationId: retrieveCategoryUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - Category
        - Product
      summary: deleteCategory
      operationId: deleteCategoryUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
    patch:
      tags:
        - Category
        - Product
      summary: updateCategory
      operationId: updateCategoryUsingPATCH
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: category
          description: category
          required: true
          schema:
            $ref: '#/definitions/Category'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
definitions:
  Category:
    type: object
    properties:
      children:
        type: array
        items:
          $ref: '#/definitions/Category'
      description:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
        description: name can not be null
      parentCategoryId:
        type: integer
        format: int64
    title: Category
  Product:
    type: object
    properties:
      categoryId:
        type: integer
        format: int64
      currency:
        type: string
        enum:
          - EUR
          - CHF
          - GBP
          - USD
          - INR
          - CNY
      description:
        type: string
      id:
        type: integer
        format: int64
      name:
        type: string
      price:
        type: number
        format: double
      priceInEuros:
        type: number
        format: double
    title: Product
  Sort:
    type: object
    properties:
      empty:
        type: boolean
      sorted:
        type: boolean
      unsorted:
        type: boolean
    title: Sort